+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * Fixed Ruby 2.0 Windows compatibility issues around ruby-wmi gem by replacing it with wmi-lite gem.
 * Set proxy environment variables if preset in config. (CHEF-4712)
 * Automatically enable verify_api_cert when running chef-client in local-mode. (Chef Issues 1464)
+* Upload potentially unsafe (not-json-friendly) registry key data to node as an md5 checksum. (CHEF-5323)
 
 ## Release: 11.12.4 (04/30/2014)
 http://www.getchef.com/blog/2014/04/30/release-chef-client-11-12-4-ohai-7-0-4/